+++ sdk/alertsmanagement/Azure.Provisioning.AlertsManagement/tests/BasicAlertsManagementTests.cs
@@ -2,8 +2,6 @@
 // Licensed under the MIT License.
 
 using System.Threading.Tasks;
-using Azure.Core;
-using Azure.Provisioning.Expressions;
 using Azure.Provisioning.Tests;
 using NUnit.Framework;
 
@@ -11,7 +9,7 @@ namespace Azure.Provisioning.AlertsManagement.Tests;
 
 public class BasicAlertsManagementTests
 {
-    internal static Trycep CreateAlertProcessingRuleTest()
+    internal static Trycep CreateServiceAlertTest()
     {
         return new Trycep().Define(
             ctx =>
@@ -19,37 +17,9 @@ public class BasicAlertsManagementTests
                 #region Snippet:AlertsManagementBasic
                 Infrastructure infra = new();
 
-                AlertProcessingRule rule =
-                    new(nameof(rule), AlertProcessingRule.ResourceVersions.V2021_08_08)
-                    {
-                        Location = new AzureLocation("global"),
-                        Properties = new AlertProcessingRuleProperties
-                        {
-                            Scopes = { BicepFunction.GetSubscription().Id },
-                            Conditions =
-                            {
-                                new AlertProcessingRuleCondition
-                                {
-                                    Field = AlertProcessingRuleField.TargetResourceType,
-                                    Operator = AlertProcessingRuleOperator.EqualsValue,
-                                    Values = { "microsoft.recoveryservices/vaults" },
-                                },
-                            },
-                            Actions =
-                            {
-                                new AlertProcessingRuleAddGroupsAction
-                                {
-                                    ActionGroupIds =
-                                    {
-                                        new ResourceIdentifier("/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/example-rg/providers/Microsoft.Insights/actionGroups/backupAlertsActionGroup"),
-                                    },
-                                },
-                            },
-                            Description = "Sample alert processing rule",
-                            IsEnabled = true,
-                        },
-                    };
-                infra.Add(rule);
+                ServiceAlert alert = ServiceAlert.FromExisting(nameof(alert), ServiceAlert.ResourceVersions.V2025_05_25_PREVIEW);
+                alert.Name = "existingAlert";
+                infra.Add(alert);
                 #endregion
 
                 return infra;
@@ -57,39 +27,13 @@ public class BasicAlertsManagementTests
     }
 
     [Test]
-    [Description("https://github.com/Azure/azure-quickstart-templates/blob/master/quickstarts/microsoft.recoveryservices/recovery-services-create-alert-processing-rule/main.bicep")]
-    public async Task CreateBackupAlertProcessingRule()
+    public async Task ReferenceExistingServiceAlert()
     {
-        await using Trycep test = CreateAlertProcessingRuleTest();
+        await using Trycep test = CreateServiceAlertTest();
         test.Compare(
             """
-            resource rule 'Microsoft.AlertsManagement/actionRules@2021-08-08' = {
-              name: take('rule${uniqueString(resourceGroup().id)}', 24)
-              location: 'global'
-              properties: {
-                scopes: [
-                  subscription().id
-                ]
-                conditions: [
-                  {
-                    field: 'TargetResourceType'
-                    operator: 'Equals'
-                    values: [
-                      'microsoft.recoveryservices/vaults'
-                    ]
-                  }
-                ]
-                actions: [
-                  {
-                    actionType: 'AddActionGroups'
-                    actionGroupIds: [
-                      '/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/example-rg/providers/Microsoft.Insights/actionGroups/backupAlertsActionGroup'
-                    ]
-                  }
-                ]
-                description: 'Sample alert processing rule'
-                enabled: true
-              }
+            resource alert 'Microsoft.AlertsManagement/alerts@2025-05-25-preview' existing = {
+              name: 'existingAlert'
             }
             """);
     }